api webhooks checked

+++ api/routes/webhook.route.js
@@ -9,7 +9,7 @@ const CalendarService = require('../services/calendar.service');
 const router = express.Router();
 
 router.post('/event', async (req, res, next) => {
-    const webhookSecret = 'your_webhook_secret';
+    const webhookSecret = 'C6nKjhS5fj9rQhcQ5hppEht44ARc2z7n559SdJt97uMZTBYJ9mejLkvGNEqXZbAAzZrc58hLQz2RQzBGePib7p5eCJaVrjysoPHAUfvvA3HJ22BZGFrf911C6skFSFYr';
     const signature = req.headers['teamup-signature'];
     const payload = JSON.stringify(req.body);
 
@@ -18,24 +18,28 @@ router.post('/event', async (req, res, next) => {
                        .digest('hex');
 
     if (hash === signature) {
-        // Do something with the event
-        console.log('Event received:', payload);
-        const trigger = payload.dispatch.trigger;
-        const eventId = payload.dispatch.event.id;
-        const event = payload.dispatch.event;
+        const dispatch = req.body.dispatch;
 
-        const subcalendar_ids = event.subcalendar_ids;
-        const employee = await DbService.getOne(COLLECTIONS.EMPLOYEES, { teamupSubCalendarId: subcalendar_ids[0] });
-        if(!employee) return res.status(404).send('Employee not found');
-        const businessId = employee.businessId;
-        const business = await DbService.getById(COLLECTIONS.BUSINESSES, { _id: businessId });
-        if(!business) return res.status(404).send('Business not found');
-        const calendar = await DbService.getOne(COLLECTIONS.CALENDARS, { businessId });
-        if(!calendar) return res.status(404).send('Calendar not found');
+        if(dispatch && dispatch.length > 0) {
+            for(let item of dispatch) {
+                const trigger = item.trigger;
+                const eventId = item.event.id;
+                const event = item.event;
 
-        switch(trigger) {
-            case 'event.created':
-                // Do something when an event is created
+                const subcalendar_ids = event.subcalendar_ids;
+                const employee = await DbService.getOne(COLLECTIONS.EMPLOYEES, { teamupSubCalendarId: subcalendar_ids[0] });
+                if(!employee) return res.status(404).send('Employee not found');
+                const businessId = employee.businessId;
+                const business = await DbService.getById(COLLECTIONS.BUSINESSES, businessId);
+                if(!business) return res.status(404).send('Business not found');
+                const calendar = await DbService.getOne(COLLECTIONS.CALENDARS, { businessId });
+                if(!calendar) return res.status(404).send('Calendar not found');
+    
+                if(trigger == 'event.removed') {
+                    await DbService.delete(COLLECTIONS.EVENTS, { teamupEventId: eventId });
+                    return res.status(200).send('Event received');
+                }
+    
                 const newEvent = new Event({
                     calendarId: new mongoose.Types.ObjectId(calendar._id),
                     teamupSubCalendarIds: event.subcalendar_ids,
@@ -45,35 +49,22 @@ router.post('/event', async (req, res, next) => {
                     start: event.start_dt,
                     end: event.end_dt
                 });
-
-                await DbService.create(COLLECTIONS.EVENTS, newEvent);
-                break;
-            case 'event.modified':
-                // Do something when an event is modified
-                if(!event.rrule) {
-                    await DbService.delete(COLLECTIONS.EVENTS, { teamupEventId: eventId });
-                    const modifiedEvent = new Event({
-                        calendarId: new mongoose.Types.ObjectId(calendar._id),
-                        teamupSubCalendarIds: event.subcalendar_ids,
-                        allDay: event.all_day,
-                        rrule: event.rrule,
-                        teamupEventId: eventId,
-                        start: event.start_dt,
-                        end: event.end_dt
-                    });
-
-                    await DbService.create(COLLECTIONS.EVENTS, modifiedEvent);
-                } else {
-                    try {
-                        await CalendarService.syncCalendar(calendar._id);
-                    } catch (error) {
-                        console.error(error);
+    
+                if(trigger == 'event.created') {
+                    await DbService.create(COLLECTIONS.EVENTS, newEvent);
+                } else if(trigger == 'event.modified') {
+                    if(!event.rrule) {
+                        await DbService.delete(COLLECTIONS.EVENTS, { teamupEventId: eventId });
+                        await DbService.create(COLLECTIONS.EVENTS, newEvent);
+                    } else {
+                        try {
+                            await CalendarService.syncCalendar(calendar._id);
+                        } catch (error) {
+                            console.error(error);
+                        }
                     }
                 }
-                break;
-            case 'event.deleted':
-                await DbService.delete(COLLECTIONS.EVENTS, { teamupEventId: eventId });
-                break;
+            }
         }
 
         res.status(200).send('Event received');
